requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
qrcode[pil]
python-dotenv
psycopg2-binary
jinja2
gunicorn
apscheduler
pytz
reportlab
pillow
